# exchange-platform/services/internal/notification-service/api/internal/api-spec.yaml
openapi: 3.0.3
info:
  title: Notification Service API
  version: 1.0.0
  description: API for the enterprise Notification Service for routing messages across multiple delivery channels.
servers:
  - url: http://internal-api.example.local/v1
    description: Internal API Gateway

paths:
  /notifications:
    post:
      x-amazon-apigateway-integration:
        type: http_proxy
        httpMethod: POST
        uri: http://notification-service.internal/notifications
        connectionType: VPC_LINK
        connectionId: "${vpc_link_id}"
      summary: Send a single notification
      description: Submit a single notification request specifying channel, destination, and template context.
      operationId: sendSingleNotification
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NotificationRequest'
      responses:
        '202':
          description: Notification conditionally accepted for processing
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationResponse'
        '400':
          description: Invalid request payload
        '500':
          description: Internal server error

  /notifications/batch:
    post:
      x-amazon-apigateway-integration:
        type: http_proxy
        httpMethod: POST
        uri: http://notification-service.internal/notifications/batch
        connectionType: VPC_LINK
        connectionId: "${vpc_link_id}"
      summary: Send multiple notifications
      description: Submit an array of notifications to be processed in a single batch.
      operationId: sendBatchNotifications
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/NotificationRequest'
      responses:
        '202':
          description: Batch accepted for processing
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BatchNotificationResponse'
        '400':
          description: Invalid request payload

  /notifications/list:
    post:
      x-amazon-apigateway-integration:
        type: http_proxy
        httpMethod: POST
        uri: http://notification-service.internal/notifications/list
        connectionType: VPC_LINK
        connectionId: "${vpc_link_id}"
      summary: Send a notification to a distribution list
      description: |
        Dispatches a notification to all active subscribers of the specified distribution list.
        The service resolves the `listId` to the underlying SFMC Data Extension and executes
        a Triggered Send for all members. Useful for broadcast or scheduled campaign delivery.
      operationId: sendToDistributionList
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ListNotificationRequest'
      responses:
        '202':
          description: Notification accepted for distribution list delivery
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationResponse'
        '400':
          description: Invalid request payload
        '404':
          description: Distribution list not found
        '500':
          description: Internal server error


  /users/{userId}/notifications:
    post:
      x-amazon-apigateway-integration:
        type: http_proxy
        httpMethod: POST
        uri: http://notification-service.internal/users/{userId}/notifications
        connectionType: VPC_LINK
        connectionId: "${vpc_link_id}"
      summary: Generate a user notification
      description: Create a notification or alert registered against an individual user.
      operationId: createUserNotification
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
          description: Unique identifier for the user.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserNotificationRequest'
      responses:
        '201':
          description: User notification created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserNotificationResponse'
        '400':
          description: Invalid request payload
        '500':
          description: Internal server error

    get:
      x-amazon-apigateway-integration:
        type: http_proxy
        httpMethod: GET
        uri: http://notification-service.internal/users/{userId}/notifications
        connectionType: VPC_LINK
        connectionId: "${vpc_link_id}"
      summary: Get user notifications
      description: Retrieve notifications and alerts for a specific user.
      operationId: getUserNotifications
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
          description: Unique identifier for the user.
      responses:
        '200':
          description: List of user notifications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserNotificationResponse'
        '500':
          description: Internal server error

  /users/{userId}/notifications/{notificationId}:
    delete:
      x-amazon-apigateway-integration:
        type: http_proxy
        httpMethod: DELETE
        uri: http://notification-service.internal/users/{userId}/notifications/{notificationId}
        connectionType: VPC_LINK
        connectionId: "${vpc_link_id}"
      summary: Delete a user notification
      description: Deletes a specific notification for a user.
      operationId: deleteUserNotification
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
        - name: notificationId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Notification deleted successfully
        '404':
          description: Notification not found
        '500':
          description: Internal server error
  /distribution-lists:
    post:
      x-amazon-apigateway-integration:
        type: http_proxy
        httpMethod: POST
        uri: http://notification-service.internal/distribution-lists
        connectionType: VPC_LINK
        connectionId: "${vpc_link_id}"
      summary: Create a distribution list
      description: Creates a new distribution list (SFMC Data Extension) for targeted notification delivery.
      operationId: createDistributionList
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DistributionListRequest'
      responses:
        '201':
          description: Distribution list created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DistributionList'
        '400':
          description: Invalid request payload
        '500':
          description: Internal server error

  /distribution-lists/{listId}:
    get:
      x-amazon-apigateway-integration:
        type: http_proxy
        httpMethod: GET
        uri: http://notification-service.internal/distribution-lists/{listId}
        connectionType: VPC_LINK
        connectionId: "${vpc_link_id}"
      summary: Get a distribution list
      description: Retrieve metadata for a specific distribution list by its identifier.
      operationId: getDistributionList
      parameters:
        - name: listId
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Unique identifier of the distribution list.
      responses:
        '200':
          description: Distribution list found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DistributionList'
        '404':
          description: Distribution list not found
        '500':
          description: Internal server error

  /distribution-lists/{listId}/subscribers:
    post:
      x-amazon-apigateway-integration:
        type: http_proxy
        httpMethod: POST
        uri: http://notification-service.internal/distribution-lists/{listId}/subscribers
        connectionType: VPC_LINK
        connectionId: "${vpc_link_id}"
      summary: Add subscribers to a distribution list
      description: Appends one or more subscribers to an existing distribution list.
      operationId: addSubscribers
      parameters:
        - name: listId
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Unique identifier of the distribution list.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/Subscriber'
      responses:
        '200':
          description: Subscribers added successfully
        '400':
          description: Invalid subscriber data
        '404':
          description: Distribution list not found
        '500':
          description: Internal server error

  /distribution-lists/{listId}/subscribers/{subscriberId}:
    delete:
      x-amazon-apigateway-integration:
        type: http_proxy
        httpMethod: DELETE
        uri: http://notification-service.internal/distribution-lists/{listId}/subscribers/{subscriberId}
        connectionType: VPC_LINK
        connectionId: "${vpc_link_id}"
      summary: Remove a subscriber from a distribution list
      description: Removes a specific subscriber from the distribution list.
      operationId: removeSubscriber
      parameters:
        - name: listId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: subscriberId
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Subscriber removed successfully
        '404':
          description: Distribution list or subscriber not found
        '500':
          description: Internal server error

components:
  schemas:
    Channel:
      type: string
      enum:
        - EMAIL
        - SMS
        - WHATSAPP
        - SIGNAL
        - BLUESKY
        - X
      description: The delivery medium for the given notification.
      
    NotificationRequest:
      type: object
      required:
        - channel
        - recipientType
        - templateId
      description: |
        Represents a single notification request. The `recipientType` field determines the
        delivery target: use `SINGLE` with a `destination` address, or `LIST` with a `listId`.
        Exactly one of `destination` or `listId` must be provided.
      properties:
        channel:
          $ref: '#/components/schemas/Channel'
        recipientType:
          type: string
          enum: [SINGLE, LIST]
          description: Determines whether to deliver to a single address or a distribution list.
        destination:
          type: string
          description: |
            The endpoint recipient (e.g., email address or phone number).
            Required when `recipientType` is `SINGLE`.
        listId:
          type: string
          format: uuid
          description: |
            The unique identifier of the distribution list to send to.
            Required when `recipientType` is `LIST`.
        source:
          type: string
          description: Originating service or user triggering the notification. Used for auditing.
        templateId:
          type: string
          description: Unique identifier of the message template to render.
        templatePayload:
          type: object
          description: Key-value map of dynamic properties injected into the template.
          additionalProperties: true

    ListNotificationRequest:
      type: object
      required:
        - channel
        - listId
        - templateId
      description: |
        A notification request targeting all members of a distribution list. The service
        resolves the `listId` to the underlying SFMC Data Extension and dispatches
        a Triggered Send to all active subscribers.
      properties:
        channel:
          $ref: '#/components/schemas/Channel'
        listId:
          type: string
          format: uuid
          description: Unique identifier of the distribution list to broadcast to.
        source:
          type: string
          description: Originating service or user triggering the notification. Used for auditing.
        templateId:
          type: string
          description: Unique identifier of the message template to render.
        templatePayload:
          type: object
          description: Key-value map of dynamic properties injected into the template.
          additionalProperties: true

    NotificationResponse:
      type: object
      properties:
        notificationId:
          type: string
          format: uuid
          description: Unique identifier for tracking this specific notification's delivery state.
        status:
          type: string
          enum: [ACCEPTED, REJECTED]
          
    BatchNotificationResponse:
      type: object
      properties:
        batchId:
          type: string
          format: uuid
        acceptedCount:
          type: integer
        notifications:
          type: array
          items:
            $ref: '#/components/schemas/NotificationResponse'

    UserNotificationRequest:
      type: object
      required:
        - message
      properties:
        message:
          type: string
          description: The main content of the notification or alert.
        metadata:
          type: object
          description: Arbitrary key-value metadata associated with the notification.
          additionalProperties: true
        activeStartDate:
          type: string
          format: date-time
          description: The date and time when the notification becomes active.
        activeEndDate:
          type: string
          format: date-time
          description: The date and time when the notification expires or becomes inactive.

    UserNotificationResponse:
      type: object
      properties:
        notificationId:
          type: string
          format: uuid
          description: Unique identifier for the user notification.
        userId:
          type: string
          description: The user this notification belongs to.
        message:
          type: string
        metadata:
          type: object
          additionalProperties: true
        activeStartDate:
          type: string
          format: date-time
        activeEndDate:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time

    DistributionListRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          description: Human-readable name for the distribution list.
        description:
          type: string
          description: Optional description of the list's purpose or audience.

    DistributionList:
      type: object
      properties:
        listId:
          type: string
          format: uuid
          description: Unique identifier for the distribution list.
        name:
          type: string
          description: Human-readable name for the distribution list.
        description:
          type: string
        externalKey:
          type: string
          description: The SFMC Data Extension key assigned by the provider upon creation.
        subscriberCount:
          type: integer
          description: Number of active subscribers in the list.
        createdAt:
          type: string
          format: date-time

    Subscriber:
      type: object
      required:
        - email
      properties:
        subscriberId:
          type: string
          description: Domain identifier for the subscriber (e.g., userId).
        email:
          type: string
          format: email
          description: Email address of the subscriber.
        firstName:
          type: string
        lastName:
          type: string
        attributes:
          type: object
          description: Additional key-value attributes to populate on the SFMC Data Extension record.
          additionalProperties: true